+++ statistics.h
@@ -7,8 +7,8 @@
 
 #endif //REGRESI_LINEAR_BERGANDA_STATISTICS_H
 
-double sum(const int numbers[], int size);
-double square_sum(const int *array, int size);
+double sum(const double numbers[], int size);
+double square_sum(const double *array, int size);
 double sigma_of_sigma(double sum_square, double sum, int count);
-double sum_multi_value(const int *array1, const int *array2, int size);
+double sum_multi_value(const double *array1, const double *array2, int size);
 double sigma_x1x2(double x1x2, double x1, double x2, int count);
